project/game_service/requirements.txt
Django
djangorestframework
PyJWT
requests
django-cors-headers
psycopg
kafka-python
channels